+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Changed `hf 14a sniff -i` - now supports the interactive flag (@iceman1001)
  - Added `HF_15SIM` standalone mode that dump then simulate iso15 tags (@lnv42)
  - Changed `iso15 simulation`, reworked, added support for lot of features (@lnv42)
  - Changed `hf 15 sniff` quality while low signal (@lnv42)